+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 1.3.0
+
+- Updated previous methods
+- Added new Methods
+  - Save Methods
+    - saveBool(String key, bool value)
+    - saveDouble(String key, double value)
+    - saveInt(String key, int value)
+    - saveString(String key, String value)
+    - saveStringList(String key, List<String> value)
+  - Get Methods
+    - getBool(String key)
+    - getDouble(String key)
+    - getInt(String key)
+    - getString(String key)
+    - getStringList(String key)
+    - getKeys()
+- Updated Documentation
+
 ## 1.2.2
 
 - Prepared The package for nullsafety